+++ src/client/mod.rs
@@ -1,8 +1,10 @@
+pub mod capture;
 pub mod connect;
 mod pull;
 mod push;
 mod watch;
 
+pub use capture::capture;
 pub use connect::status;
 pub use pull::pull;
 pub use push::push;